64-bit fixes for sipXtackLib. Use intptr_t, UtlVoidPtr, etc when appropriate.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@10912 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXtackLib/include/net/SipUdpServer.h
@@ -129,8 +129,8 @@ public:
 
     void sendSipKeepAlive(OsTimer* pTimer) ; 
 
-    static void SipKeepAliveCallback(const int userData, 
-                                     const int eventData) ;
+    static void SipKeepAliveCallback(const intptr_t userData, 
+                                     const intptr_t eventData) ;
 
 /* ============================ ACCESSORS ================================= */
